+++ sync-for-payables/CodatSyncPayables/Models/Operations/DeleteBillPaymentRequest.cs
@@ -8,9 +8,9 @@
 // </auto-generated>
 //------------------------------------------------------------------------------
 #nullable enable
-namespace CodatSyncPayables.Models.Operations
+namespace Codat.Sync.Payables.Models.Operations
 {
-    using CodatSyncPayables.Utils;
+    using Codat.Sync.Payables.Utils;
     
     public class DeleteBillPaymentRequest
     {
@@ -26,5 +26,11 @@ namespace CodatSyncPayables.Models.Operations
         /// </summary>
         [SpeakeasyMetadata("pathParam:style=simple,explode=false,name=companyId")]
         public string CompanyId { get; set; } = default!;
+
+        /// <summary>
+        /// Unique identifier for a connection.
+        /// </summary>
+        [SpeakeasyMetadata("pathParam:style=simple,explode=false,name=connectionId")]
+        public string ConnectionId { get; set; } = default!;
     }
 }